Migrated http function to be callable by the sdk
