+++ functions/package.json
@@ -14,22 +14,16 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "@google-cloud/storage": "^4.7.0",
-    "camelcase": "^5.3.1",
-    "cookie-parser": "^1.4.4",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
     "firebase-admin": "^8.12.1",
-    "firebase-function-tools": "^1.1.3",
     "firebase-functions": "^3.6.2",
-    "fs-extra": "^8.1.0",
+    "firebase-functions-test": "^0.2.0",
+    "firebase-function-tools": "^1.1.3",
     "glob": "^7.1.6",
-    "uuid": "^7.0.3"
+    "camelcase": "^5.3.1"
   },
   "devDependencies": {
     "tslint": "^5.12.0",
-    "typescript": "^3.8.0",
-    "firebase-functions-test": "^0.2.0"
+    "typescript": "^3.8.0"
   },
   "private": true
 }
