--- functions/src/requests/middlewares/auth.ts
@@ -1,62 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-import * as admin from 'firebase-admin';
-
-try {
-  admin.initializeApp();
-} catch (e) {}
-
-const validateFirebaseIdToken = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  console.log('Check if request is authorized with Firebase ID token');
-
-  if (
-    (!req.headers.authorization ||
-      !req.headers.authorization.startsWith('Bearer ')) &&
-    !(req.cookies && req.cookies.__session)
-  ) {
-    console.error(
-      'No Firebase ID token was passed as a Bearer token in the Authorization header.',
-      'Make sure you authorize your request by providing the following HTTP header:',
-      'Authorization: Bearer <Firebase ID Token>',
-      'or by passing a "__session" cookie.'
-    );
-    res.status(403).send('Unauthorized');
-    return;
-  }
-
-  let idToken;
-  if (
-    req.headers.authorization &&
-    req.headers.authorization.startsWith('Bearer ')
-  ) {
-    console.log('Found "Authorization" header');
-    // Read the ID Token from the Authorization header.
-    idToken = req.headers.authorization.split('Bearer ')[1];
-  } else if (req.cookies) {
-    console.log('Found "__session" cookie');
-    // Read the ID Token from cookie.
-    idToken = req.cookies.__session;
-  } else {
-    // No cookie
-    res.status(403).send('Unauthorized');
-    return;
-  }
-
-  try {
-    const decodedIdToken = await admin.auth().verifyIdToken(idToken);
-    console.log('ID Token correctly decoded', decodedIdToken);
-    // @ts-ignore
-    req.user = decodedIdToken;
-    next();
-    return;
-  } catch (error) {
-    console.error('Error while verifying Firebase ID token:', error);
-    res.status(403).send('Unauthorized');
-    return;
-  }
-};
-
-export default validateFirebaseIdToken;
