+++ src/firebase.js
@@ -2,6 +2,7 @@ import firebase from 'firebase/app';
 import 'firebase/database';
 import 'firebase/auth';
 import 'firebase/storage';
+import 'firebase/functions';
 
 const config = {
   apiKey: process.env.REACT_APP_FIRE_BASE_KEY,
